modify js with modules
