+++ index.html
@@ -173,7 +173,7 @@
         <div class="cards-container"></div>
     </section>
 
-    <script src="./products.js"></script>
+    <script src="./main.js" type="module"></script>
 </body>
 
 </html>
